refactor route and fix pinia
